+++ utah-design-system-website/src/react/propTypesShapes/TooltipsExamplePropsShape.js
@@ -1,5 +1,11 @@
 import PropTypes from 'prop-types';
 
 export default PropTypes.exact({
-  onClick: PropTypes.func,
+  isPopperVisible: PropTypes.bool,
+
+  // https://popper.js.org/docs/v2/modifiers/offset/
+  offsetDistance: PropTypes.string,
+  offsetSkidding: PropTypes.string,
+
+  popupText: PropTypes.string,
 });