+++ CHANGELOG.md
@@ -8,11 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## Added
-- ToolTip Component
+- Tooltip Component
 
 ## Changed
 - Change Utah Header wrapper from a `<div>` to a `<header>` element.
-- Have Icon Button show a ToolTip 
+- Have Icon Button show a Tooltip
+- Show title on icon button
 
 ## Deprecated